requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
PyJWT==2.8.0
cryptography==41.0.7
python-decouple==3.8
requests==2.31.0
celery==5.3.4
django-celery-beat==2.5.0
drf-spectacular==0.27.0
django-ratelimit==4.1.0
Pillow==10.1.0
django-extensions==3.2.3
ipython==8.17.2